samples: fwupdate: Add sample for firmware update
This sample is using SD card to do fwupdate.

If SD card contains 'SD/job.txt' in the root of directory,
and the file contains 'FWUPDATE'. Then the sample will
read firmware from 'SD/fw.elf' and burn it into secondary
partition of flash.

The partition manager is based on u-boot. And the sample
will update u-boot env to indicate which partiton needs
to be booted.

This sample also implemented fallback mechanism.

Signed-off-by: Savent Gate <savent_gate@outlook.com>