fix: gmac: Fix incorrect PHY configuration issues
Modify the PHY configuration handling method to ensure proper
operation with all MDIO  addresses, not just specific ones.
Three "partner" fields are added to the device tree, which enables
eth0 to easily obtain the attributes of eth1. This ensures that all PHY
parameters in the device tree are correctly applied.

Signed-off-by: Gravity Constant <constants667@outlook.com>